[Cleanup] Extract shared SinusPositionEmbedding
Move the duplicated SinusPositionEmbedding class (sinusoidal timestep
conditioning for DiT/CFM) into a shared module at
vllm_omni/model_executor/layers/timestep_embedding.py.

Replaces 4 identical/near-identical copies:
- qwen3_tts/tokenizer_25hz/modeling_qwen3_tts_tokenizer_v1.py
- qwen2_5_omni/qwen2_5_omni_token2wav.py
- ming_flash_omni/talker_module.py
- cosyvoice3_audio/cosyvoice3_dit.py
